Pull out source counting to new file
Fix some issues with glossaries; will be used in future changes
