+++ ApproachGlossary.csv
@@ -301,7 +301,7 @@ Neuron Coverage (Testing)?,Approach,"Testing based on ""the coverage of activate
 "Non-functional Testing (Washizaki, 2024, p. 5-8)",Approach,"Testing that ""targets the validation of non-functional aspects (such as performance, [security,] usability, or reliability)"" (Washizaki, 2024, p. 5-8; IEEE, 2022, p. 21; similar in Hamburg and Mogyorodi, 2024) that tend to ""constrain the technologies to be used in the implementation"" (Washizaki, 2024, p. 1-4)","Requirements-based Testing (Kam, 2008, p. 47), Dynamic Testing (although this seems over-zealous), W-Model Testing (Gerrard, 2000a, p. 9)",,"Hundreds of subapproaches that can be ""performed at all test levels"" (Washizaki, 2024, p. 5-8). ""Nonfunctional requirements"" are also called ""performance attributes"" (IEEE, 2017, p. 293); what does this imply for the relationship between non-functional and performance testing? In constrast to functional testing (IEEE, 2022, p. 21)"
 N-Switch Testing,"Technique (inferred from state transition testing and IEEE, 2021, p. 20)","Testing ""derived to cover valid sequences of N + 1 transitions in the state model"" (IEEE, 2021, p. 20)","State Transition Testing (IEEE, 2021, p. 20; Kam, 2008, p. 46)","Multiple Transitions Testing (implied by IEEE, 2021, p. 20)","See Chow, 1978"
 Object-based Testing,Approach,,"Web Application Testing, Object-Oriented Testing, Structure-based Testing, Data Flow Testing, Functional Testing (Kam, 2008, Tab. 1)",,
-"Object-Oriented Testing (Doğan et al., 2014, Tab. 1)",Practice?,,,,See Binder (1996)
+"Object-Oriented Testing (Doğan et al., 2014, Tab. 1)",Practice?,,,,"See Binder, 1996"
 Offline MBT,"Approach (Hamburg and Mogyorodi, 2024)","""Model-based test[ing] … whereby test cases are generated into a repository for future execution"" (Hamburg and Mogyorodi, 2024)","Offline Testing, Model-based Testing (Hamburg and Mogyorodi, 2024)",,
 Offline Testing,Practice?,"Testing ""in an environment without external interaction"" (Washizaki, 2024, p. 5-6), or testing ""whereby test cases are generated into a repository for future execution"" (Hamburg and Mogyorodi, 2024)",,,"Sometimes spelled with a hyphen (""off-line""; Washizaki, 2024, p. 5-6)"
 One-to-One Testing?,Technique?,"Testing where ""each test case is derived to exercise a specific"" test coverage item, such as an ""option or mutation"" for syntax testing (IEEE, 2021, p. 15) or equivalence partition (p. 11; OG BS 7925-2; Myers 1979)",,,
