+++ SuppGlossary.csv
@@ -7,7 +7,7 @@ CFG,,,
 Classification Tree,"""A tree diagram representing test data domains of a test object"" (Hamburg and Mogyorodi, 2024) as ""a decomposition hierarchy"" (IEEE, 2017, p. 485; OG PMBOK) using ""distinct classifications (relevant test aspects) and classes (input values)"" (IEEE, 2022, p. 3; 2021, p. 2)",,
 Commercial Off-the-Shelf,"""A type of product developed in an identical format for a large number of customers in the general market"" (Hamburg and Mogyorodi, 2024)","Off-the-Shelf Software (Hamburg and Mogyorodi, 2024)",
 Common Gateway Interface,"An ""object[] that execute[s] on the server … initiated by [a] forms-based user interaction[] on the browser"" (Gerrard, 2000b, p. 13)","CGI (Gerrard, 2000a, p. 17; 2000b, p. 11)","""The most common mechanism used to implement functionality on web sites"" (Gerrard, 2000b, p. 11), including business logic, order or payment processing, and security checking (p. 13)"
-(Software) Component,"A discrete entity, either structurally (IEEE, 2017, p. 419; OG ISO/IEC, 2015), functionally, or logically (IEEE, 2017, p. 419), ""within a system considered at a particular level of analysis"" (ISO/IEC, 2023; IEEE, 2017, p. 419; OG ISO/IEC, 2015) ""that can be tested in isolation"" (Hamburg and Mogyorodi, 2024); ""the individual subprograms or components, a subsystem, or a composition of … components"" (Washizaki, 2024, p. 5-6) or even an entire system (IEEE, 2017, p. 419; OG ISO/IEC, 2016)","Unit, often (IEEE, 2017, p. 82)",
+(Software) Component,"A discrete entity, either structurally (IEEE, 2017, p. 419; OG ISO/IEC, 2015), functionally, or logically (IEEE, 2017, p. 419), ""within a system considered at a particular level of analysis"" (ISO/IEC, 2023b; IEEE, 2017, p. 419; OG ISO/IEC, 2015) ""that can be tested in isolation"" (Hamburg and Mogyorodi, 2024); ""the individual subprograms or components, a subsystem, or a composition of … components"" (Washizaki, 2024, p. 5-6) or even an entire system (IEEE, 2017, p. 419; OG ISO/IEC, 2016)","Unit, often (IEEE, 2017, p. 82)",
 Computation Data Use,"A ""use of the value of a variable in any type of statement"" (IEEE, 2021, p. 2; 2017, p. 83; OG 2015), ""as an input to the computation of the definition of any variable or of an output"" (2021, p. 27; similar in van Vliet, 2000, p. 424), or not as a P-use; e.g., the use of a variable in a computation or I/O statement (van Vliet, 2000, p. 424)","C-use, c-use",
 Condition,"A ""Boolean expression containing no Boolean operators"" (IEEE, 2021, p. 2)","Atomic Condition (Hamburg and Mogyorodi, 2024)",
 Data Analysis,"The ""systematic investigation of the data and their flow in a real or planned system"" (IEEE, 2017, p. 114)",,
